Implement naive HTTP client based on cURL
The http module implements simple requests based on libcurl.
The requests can be get, post, put, delete or patch. Optionally,
it can be defined with content-type header and a body.

The module has been tested with the main application and using
the https://beeceptor.com/ service to analize the requests and
responses.