Simply posting data to HTTP
After a bunch of hours lost in the libcurl documentation, finally
I have undertand what simply is sending data with curl.

When the content is string, we does not have a "read_callback" to
read the body buffer to write to request. Just a zero-ending
string and set the option to POST data. The option to POST set
the content type to url-encoded automaticaly, but the code overwrites
this header.

Maybe for more complex request's body the READFUNCTION is needed. But
for this simple case to post JSONs, this implementation is enough.