Move CLI commands to the cli project
Despite the args.h has a very handy function to parse the commands
and it can be tested in the test project, I think that the
commands should be implemented in the cli project itself.

This commit moves the commands from args.h to cfcli. Also implement
a new one for "help" and prints the usage of the tool.